Pcla tutorial fixes (#5313)
* fixes


* fixes


* moved `create_text_and_labels` to token_classification_utils.py
